// app/booru/szbooru/types/api/post.ts
// Generated by https://quicktype.io

export interface SZ_TYPE_Post {
    id:                 number;
    version:            number;
    creationTime:       Date;
    lastEditTime:       Date;
    safety:             string;
    source:             null;
    type:               string;
    mimeType:           string;
    checksum:           string;
    checksumMD5:        string;
    fileSize:           number;
    canvasWidth:        number;
    canvasHeight:       number;
    contentUrl:         string;
    thumbnailUrl:       string;
    flags:              any[];
    tags:               SZ_POST_Tag[];
    relations:          any[];
    user:               SZ_POST_User;
    score:              number;
    ownScore:           number;
    ownFavorite:        boolean;
    tagCount:           number;
    favoriteCount:      number;
    commentCount:       number;
    noteCount:          number;
    relationCount:      number;
    featureCount:       number;
    lastFeatureTime:    null;
    favoritedBy:        any[];
    hasCustomThumbnail: boolean;
    notes:              any[];
    comments:           any[];
    pools:              any[];
}

export interface SZ_POST_Tag {
    names:    string[];
    category: string;
    usages:   number;
}

export interface SZ_POST_User {
    name:      string;
    avatarUrl: string;
}